refactor : `LadderGenerator` -> `Ladder` 네이밍 변경
(cherry picked from commit f25fe94e3ebb89bd80b1bf0d0b1db30cddc31b82)
