--- src/test/java/domain/LadderGeneratorTest.java
@@ -1,21 +1,0 @@
-package domain;
-
-import org.assertj.core.api.Assertions;
-import org.junit.jupiter.api.Test;
-
-import java.util.Arrays;
-import java.util.LinkedList;
-import java.util.Queue;
-
-class LadderGeneratorTest {
-
-    @Test
-    void getLadder() {
-        Queue<Boolean> randomNumber = new LinkedList<>();
-        Arrays.asList(true, true, false).forEach(randomNumber::add);
-        LadderGenerator ladderGenerator = new LadderGenerator(5, new CustomRandomGenerator(randomNumber));
-
-        Assertions.assertThat(ladderGenerator.getLadder())
-                .isEqualTo(Arrays.asList(Position.LEFT, Position.RIGHT, Position.LEFT, Position.RIGHT, Position.DOWN));
-    }
-}
